pyspark-code/requirements.txt
pyspark
pymongo
requests
py7zr
beautifulsoup4
python-dotenv
pandas